Добавил проверку ERROR на сайте gta-objects. Если img_url = None то программа пропускает элемент и работает со следующим

+++ main.py
@@ -7,6 +7,7 @@ from chromedriver import driver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+from selenium.common.exceptions import NoSuchElementException, TimeoutException
 
 # Запрашиваем у пользователя выбор сайта для работы и преобразуем введенное значение в целое число
 user_check_site = int(input('Укажите сайт, через который вы хотите работать, 1 - plebmasters; 2 - gta-objects: '))
@@ -50,18 +51,31 @@ def plebmasters(input_object):
 
 # Функция сайта gta-objects, производящая поиск изображения объекта на сайте, подставляя hash объекта в адресную строку
 def gta_objects_xyz(input_object):
+    img_url = f'https://gta-objects.xyz/gallery/objects/{input_object}.jpg'
+    print(f'Сейчас обрабатывается: {input_object}')
+    print(f"Потенциальное URL изображения: {img_url}")
+
+    # Проверка доступности hash'а объекта на сайте
     try:
-        img_url = f'https://gta-objects.xyz/gallery/objects/{input_object}.jpg'
-        print(f'Сейчас обрабатывается: {input_object}')
-        print(f"URL изображения: {img_url}")
+        # Открываем URL, указанный в img_url. Ожидаем, пока 'title' станет видимым на странице.
+        driver.get(img_url)
+        wait.until(EC.visibility_of_element_located((By.CLASS_NAME, 'title')))
 
-        # Возвращаем переменную, для последующей работы другой функцию
+        # Находим элемент с заголовком ошибки
+        error_element = driver.find_element(By.XPATH, '//h1[@class="title pt-4 pb-4"]')
+
+        # Проверяем, содержит ли текст элемента сообщение об ошибке 404.
+        # Если ошибка найдена, выводим сообщение об ошибке и возвращаем None
+        if "ERROR 404 :: Page not found" in error_element.text:
+            print(f"Объект {input_object} не найден: ERROR 404")
+            return None
+        else:
+            # Если ошибка 404 не найдена, возвращаем URL изображения
+            return img_url
+
+    # Обработка исключений, если элемент не найден или истекло время ожидания то возвращаем img_url
+    except (NoSuchElementException, TimeoutException):
         return img_url
-
-    # Обрабатываем возможные ошибки
-    except Exception as e:
-        print(f"Ошибка при обработке {input_object} на gta_objects_xyz: объект не найден. {e}")
-        return None
 
 
 # Главная функция программы
@@ -71,39 +85,41 @@ def main(object_list):
 
     # Итерация по каждому объекту в списке object_list
     for input_object in object_list:
-        try:
-            # Проверка выбора сайта пользователем и вызов соответствующей функции с записей результата в переменную
-            if user_check_site == 1:
-                img_url = plebmasters(input_object)
-            else:
-                img_url = gta_objects_xyz(input_object)
+        # Проверка выбора сайта пользователем и вызов соответствующей функции с записей результата в переменную
+        if user_check_site == 1:
+            img_url = plebmasters(input_object)
+        else:
+            img_url = gta_objects_xyz(input_object)
 
-            # Запись описание изображения в переменную "description"
-            description = visionbot(img_url)
-
-            # Очищение файлов куки, сессии и локальных файлов сайта, для обеспечения лучшей работы программы
-            driver.delete_all_cookies()
-            driver.execute_script("window.localStorage.clear();")
-            driver.execute_script("window.sessionStorage.clear();")
-
-            # Запись сокращенного описания в переменную "short_description"
-            short_description = description_image(description)
-
-            # Запись hash'а и его сокращенного описания в список
-            results.append({
-                'hash': input_object,
-                'Description': short_description
-            })
-
-        # Обрабатываем возможные ошибки
-        except Exception as e:
-            print(f"Ошибка при обработке объекта {input_object}: {e}")
+        # Проверка является ли переменная img_url = None
+        if img_url is None:
+            print(f"Пропущен объект {input_object}: получен None для URL изображения.")
 
             # Запись hash'а с сообщением об ошибке выполнения
             results.append({
                 'hash': input_object,
-                'Description': "Ошибка выполнения"
+                'Description': "Файл не найден."
             })
+
+            # Продолжение программы со следующим элементом списка, пропуская ошибочный.
+            continue
+
+        # Запись описание изображения в переменную "description"
+        description = visionbot(img_url)
+
+        # Очищение файлов куки, сессии и локальных файлов сайта, для обеспечения лучшей работы программы
+        driver.delete_all_cookies()
+        driver.execute_script("window.localStorage.clear();")
+        driver.execute_script("window.sessionStorage.clear();")
+
+        # Запись сокращенного описания в переменную "short_description"
+        short_description = description_image(description)
+
+        # Запись hash'а и его сокращенного описания в список
+        results.append({
+            'hash': input_object,
+            'Description': short_description
+        })
 
     # Создание DataFrame с использованием списка и сохранением его в CSV формат без записи индексов строк
     df = pd.DataFrame(results)
